+++ qualang_tools/results/data_handler/data_processors/__init__.py
@@ -6,7 +6,7 @@ from .numpy_array_saver import NumpyArraySaver
 from .xarray_saver import XarraySaver
 from .waveform_report_saver import WaveformReportSaver
 from .simulator_controller_samples_saver import SimulatorControllerSamplesSaver
-
+from .plotly_graph_saver import PlotlyGraphSaver
 
 DEFAULT_DATA_PROCESSORS: List[Type[DataProcessor]] = [
     DataProcessor,
@@ -20,6 +20,13 @@ try:
     import xarray  # noqa: F401
 
     DEFAULT_DATA_PROCESSORS.append(XarraySaver)
+except ImportError:
+    pass
+
+try:
+    import plotly  # noqa: F401
+
+    DEFAULT_DATA_PROCESSORS.append(PlotlyGraphSaver)
 except ImportError:
     pass