+++ tests/data_handler/test_xarray_saver.py
@@ -1,37 +1,7 @@
 import copy
 import pytest
-import numpy as np
-import xarray as xr
+from test_utils import module_installed, dicts_equal
 from qualang_tools.results.data_handler.data_processors import XarraySaver
-
-
-def module_installed(module_name):
-    try:
-        exec(f"import {module_name}")
-    except ImportError:
-        return False
-    return True
-
-
-def dicts_equal(d1, d2):
-    if d1.keys() != d2.keys():
-        return False
-    for key in d1:
-        if key not in d2:
-            return False
-        elif isinstance(d1[key], dict):
-            if not dicts_equal(d1[key], d2[key]):
-                return False
-        elif isinstance(d1[key], np.ndarray):
-            if not np.array_equal(d1[key], d2[key]):
-                return False
-        elif isinstance(d1[key], xr.Dataset):
-            if not d1[key].identical(d2[key]):
-                return False
-        else:
-            if not bool(d1[key] == d2[key]):
-                return False
-    return True
 
 
 @pytest.mark.skipif(not module_installed("xarray"), reason="xarray not installed")
@@ -60,7 +30,10 @@ def test_xarray_data_saver_suffixes():
     assert xarray_saver.file_suffix == ".zarr"
 
 
-@pytest.mark.skipif(not (module_installed("xarray") and module_installed("netCDF4")), reason="xarray not installed")
+@pytest.mark.skipif(
+    not (module_installed("xarray") and module_installed("netCDF4")),
+    reason="xarray not installed",
+)
 def test_xarray_saver_merge_netcdf(tmp_path):
     import xarray as xr
 
@@ -69,7 +42,12 @@ def test_xarray_saver_merge_netcdf(tmp_path):
     xarray_saver = XarraySaver(merge_arrays=True, file_format="nc")
     processed_data = xarray_saver.process(data)
 
-    assert processed_data == {"a": 1, "b": 2, "c": "./xarrays.nc#c", "d": "./xarrays.nc#d"}
+    assert processed_data == {
+        "a": 1,
+        "b": 2,
+        "c": "./xarrays.nc#c",
+        "d": "./xarrays.nc#d",
+    }
 
     xarray_saver.post_process(data_folder=tmp_path)
 
@@ -79,7 +57,10 @@ def test_xarray_saver_merge_netcdf(tmp_path):
     xr.load_dataset(tmp_path / "xarrays.nc", group="d")
 
 
-@pytest.mark.skipif(not (module_installed("xarray") and module_installed("netCDF4")), reason="xarray not installed")
+@pytest.mark.skipif(
+    not (module_installed("xarray") and module_installed("netCDF4")),
+    reason="xarray not installed",
+)
 def test_xarray_saver_merge_hdf5(tmp_path):
     import xarray as xr
 
@@ -88,7 +69,12 @@ def test_xarray_saver_merge_hdf5(tmp_path):
     xarray_saver = XarraySaver(merge_arrays=True, file_format="h5")
     processed_data = xarray_saver.process(data)
 
-    assert processed_data == {"a": 1, "b": 2, "c": "./xarrays.h5#c", "d": "./xarrays.h5#d"}
+    assert processed_data == {
+        "a": 1,
+        "b": 2,
+        "c": "./xarrays.h5#c",
+        "d": "./xarrays.h5#d",
+    }
 
     xarray_saver.post_process(data_folder=tmp_path)
 
@@ -138,7 +124,10 @@ def test_xarray_saver_no_merge_hdf5_nested(tmp_path):
     xr.load_dataset(tmp_path / "d.d1.nc")
 
 
-@pytest.mark.skipif(not (module_installed("xarray") and module_installed("netCDF4")), reason="xarray not installed")
+@pytest.mark.skipif(
+    not (module_installed("xarray") and module_installed("netCDF4")),
+    reason="xarray not installed",
+)
 def test_xarray_saver_merge_hdf5_nested(tmp_path):
     import xarray as xr
 
@@ -147,7 +136,12 @@ def test_xarray_saver_merge_hdf5_nested(tmp_path):
     xarray_saver = XarraySaver(merge_arrays=True, file_format="nc")
     processed_data = xarray_saver.process(data)
 
-    assert processed_data == {"a": 1, "b": 2, "c": "./xarrays.nc#c", "d": {"d1": "./xarrays.nc#d.d1"}}
+    assert processed_data == {
+        "a": 1,
+        "b": 2,
+        "c": "./xarrays.nc#c",
+        "d": {"d1": "./xarrays.nc#d.d1"},
+    }
     xarray_saver.post_process(data_folder=tmp_path)
 
     assert (tmp_path / "xarrays.nc").exists()